@@ -11,21 +11,20 @@
 
     <properties>
         <!--<skipTests>true</skipTests>-->
-        <maven.compiler.target>1.8</maven.compiler.target>
-        <maven.compiler.source>1.8</maven.compiler.source>
+        <maven.compiler.target>21</maven.compiler.target>
+        <maven.compiler.source>21</maven.compiler.source>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
 
-        <mybatis-starter.version>2.1.4</mybatis-starter.version>
-        <spring-boot.version>2.3.4.RELEASE</spring-boot.version>
-        <h2.version>1.4.200</h2.version>
+        <mybatis-starter.version>3.0.2</mybatis-starter.version>
+        <spring-boot.version>3.1.0</spring-boot.version>
         <cola.components.version>4.0.0</cola.components.version>
     </properties>
 
     <modules>
         <module>think-app</module>
         <module>think-client</module>
-        <module>think-controller</module>
+        <module>think-adapter</module>
         <module>think-domain</module>
         <module>think-infrastructure</module>
         <module>think-start</module>
