@@ -1,7 +1,7 @@
 package com.amos.think.user;
 
 import com.amos.think.dto.UserRegisterCmd;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 public class UserValidatorTest {
 
