@@ -1,15 +1,15 @@
 package com.amos.think.user.command.query;
 
 import com.alibaba.cola.dto.SingleResponse;
-import com.amos.think.common.exception.BizException;
+import com.amos.think.common.exception.ThinkBizException;
 import com.amos.think.domain.gateway.UserGateway;
 import com.amos.think.domain.user.UserEntity;
 import com.amos.think.dto.data.ErrorCode;
 import com.amos.think.dto.data.UserVO;
 import com.amos.think.user.assembler.UserAssembler;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-import javax.annotation.Resource;
 import java.util.Objects;
 
 /**
@@ -21,13 +21,13 @@ import java.util.Objects;
 @Component
 public class UserInfoQueryExe {
 
-    @Resource
+    @Autowired
     private UserGateway userGateway;
 
     public SingleResponse<UserVO> execute(Long id) {
         UserEntity userEntity = userGateway.findById(id);
         if (Objects.isNull(userEntity)) {
-            throw new BizException(ErrorCode.B_USER_UNDEFINED);
+            throw new ThinkBizException(ErrorCode.B_USER_UNDEFINED);
         }
 
         return SingleResponse.of(UserAssembler.toValueObject(userEntity));
