@@ -1,13 +1,13 @@
 package com.amos.think.user.command.query;
 
-import com.amos.think.common.exception.BizException;
+import com.amos.think.common.exception.ThinkBizException;
 import com.amos.think.domain.gateway.UserGateway;
 import com.amos.think.domain.user.UserEntity;
 import com.amos.think.dto.data.ErrorCode;
 import com.amos.think.dto.query.UserLoginQuery;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-import javax.annotation.Resource;
 import java.util.Objects;
 
 /**
@@ -19,18 +19,18 @@ import java.util.Objects;
 @Component
 public class UserLoginQueryExe {
 
-    @Resource
+    @Autowired
     private UserGateway userGateway;
 
     public void execute(UserLoginQuery query) {
         UserEntity userEntity = userGateway.findPasswordInfo(query.getUsername());
         if (Objects.isNull(userEntity)) {
-            throw new BizException(ErrorCode.B_USER_PASSWORD_ERROR);
+            throw new ThinkBizException(ErrorCode.B_USER_PASSWORD_ERROR);
         }
 
         // 校验密码是否正确
         if (!userEntity.getPassword().isCorrect(query.getPassword())) {
-            throw new BizException(ErrorCode.B_USER_PASSWORD_ERROR);
+            throw new ThinkBizException(ErrorCode.B_USER_PASSWORD_ERROR);
         }
     }
 
