@@ -1,6 +1,6 @@
 package com.amos.think.gateway.impl;
 
-import com.amos.think.common.exception.BizException;
+import com.amos.think.common.exception.ThinkBizException;
 import com.amos.think.convertor.UserConvertor;
 import com.amos.think.domain.gateway.UserGateway;
 import com.amos.think.domain.user.UserEntity;
@@ -13,10 +13,10 @@ import com.amos.think.gateway.impl.database.mapper.UserInfoMapper;
 import com.amos.think.gateway.impl.database.mapper.UserMapper;
 import org.apache.commons.lang3.tuple.ImmutablePair;
 import org.apache.ibatis.exceptions.PersistenceException;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
-import javax.annotation.Resource;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -32,9 +32,9 @@ import java.util.Optional;
 @Component("userGateway")
 public class UserGatewayImpl implements UserGateway {
 
-    @Resource
+    @Autowired
     private UserMapper userMapper;
-    @Resource
+    @Autowired
     private UserInfoMapper userInfoMapper;
 
     @Override
@@ -52,8 +52,8 @@ public class UserGatewayImpl implements UserGateway {
     @Override
     public UserEntity findById(Long id) {
         Optional<UserDO> findById = userMapper.selectById(id);
-        if (!findById.isPresent()) {
-            throw new BizException(ErrorCode.B_USER_UNDEFINED);
+        if (findById.isEmpty()) {
+            throw new ThinkBizException(ErrorCode.B_USER_UNDEFINED);
         }
 
         UserDO userDO = findById.get();
@@ -119,8 +119,8 @@ public class UserGatewayImpl implements UserGateway {
      */
     private UserEntity modifyUser(UserEntity userEntity) {
         Optional<UserDO> findById = userMapper.selectById(userEntity.getId());
-        if (!findById.isPresent()) {
-            throw new BizException(ErrorCode.B_USER_UNDEFINED);
+        if (findById.isEmpty()) {
+            throw new ThinkBizException(ErrorCode.B_USER_UNDEFINED);
         }
 
         UserDO userDO = findById.get();
@@ -136,7 +136,7 @@ public class UserGatewayImpl implements UserGateway {
         }
 
         // 2. 再保存userDO
-        userDO.setGmtModify(LocalDateTime.now());
+        userDO.setGmtModified(LocalDateTime.now());
         update = userMapper.update(userDO);
         if (update < 1) {
             throw new PersistenceException("更新用户异常");
